OMAP4: PM: VC: allow channels use of default channel i2c_slaveaddr
OMAP4's PRM_VC_CFG_CHANNEL register allows for flexibility of configuring
for various PMIC configurations. In combinations where the same slave address
is used for all domains, it is possible to setup the VC channel for the
dependent channels to use the same slave address as the default channel.

Since I2C addressing could be 7 bit or 11 bits as per the I2C specification,
we use the BIT(15) to flag that this should use the default channel's
configuration. Depending on the PMIC and platform used, this can
be populated on the PMIC's datastructure and percolates to VC's configuration.

+++ arch/arm/mach-omap2/vc.h
@@ -77,7 +77,7 @@ struct omap_vc_channel {
 	u8 flags;
 
 	/* channel state */
-	u8 i2c_slave_addr;
+	u16 i2c_slave_addr;
 	u8 volt_reg_addr;
 	u8 cmd_reg_addr;
 	u8 cfg_channel;
